refactor: #31 - StudyInteraction 책임 분리
