+++ src/test/http/study/StudyInteraction.http
@@ -41,7 +41,7 @@ Content-Type: application/json
 
 
 ### 스터디 랭킹 보드 조회 (일간 순위)
-GET http://localhost:8080/api/studies/3/rankings?type=today
+GET http://localhost:8080/api/studies/1/rankings?type=today
 Content-Type: application/json
 
 ### 스터디 랭킹 보드 조회 (전체 순위)
@@ -49,7 +49,7 @@ GET http://localhost:8080/api/studies/3/rankings?type=total
 Content-Type: application/json
 
 ### 인기 스터디 목록 조회
-GET http://localhost:8080/api/studies/hit?page=1
+GET http://localhost:8080/api/studies/hit?page=0
 Content-Type: application/json
 
 ### 스터디 즐겨찾기 / 취소
